[refactor] Moved some headers and cleaned up includes
