+++ apps/pallas_info.cpp
@@ -5,17 +5,16 @@
 #include <cstdlib>
 #include <cstring>
 #include <iomanip>
+#include <iostream>
 #if __GNUC__ >= 13 || __clang__ >= 14 || _MSC_VER >= 1929
 #include <format>
 #define HAS_FORMAT
 #endif
-#include <pallas/pallas_parameter_handler.h>
 
 #include "pallas/pallas.h"
 #include "pallas/pallas_archive.h"
-#include "pallas/pallas_log.h"
-#include "pallas/pallas_read.h"
-#include "pallas/pallas_storage.h"
+#include "pallas/utils/pallas_parameter_handler.h"
+#include "pallas/utils/pallas_storage.h"
 
 #define DURATION_WIDTH 15
