+++ libraries/otf2/include/otf2/otf2.h
@@ -22,7 +22,6 @@ extern "C++" {
 #ifndef __otf2_deprecated__
 #define __otf2_deprecated__
 #endif
-#include "pallas/pallas_log.h"
 #include "pallas/pallas.h"
 #include "pallas/pallas_archive.h"
 #include "pallas/pallas_write.h"
