+++ libraries/pallas/include/pallas/pallas_read.h
@@ -11,7 +11,7 @@
 #endif
 #include "pallas.h"
 #include "pallas_attribute.h"
-#include "pallas_timestamp.h"
+#include "utils/pallas_timestamp.h"
 
 #ifdef __cplusplus
 namespace pallas {
