+++ libraries/pallas/include/pallas/pallas_record.h
@@ -9,7 +9,6 @@
 #include <stdint.h>
 #endif
 #include "pallas.h"
-#include "pallas_archive.h"
 #include "pallas_attribute.h"
 #include "pallas_write.h"
 #include "pallas_read.h"
