+++ libraries/pallas/include/pallas/utils/pallas_timestamp.h
@@ -7,12 +7,9 @@
  */
 #pragma once
 
-#include "pallas/pallas_dbg.h"
-
 #ifdef __cplusplus
 #include <chrono>
 #include <cstdint>
-#include <limits>
 typedef std::chrono::time_point<std::chrono::high_resolution_clock> Timepoint;
 #else
 #include <limits.h>
