+++ libraries/pallas/src/pallas_attribute.cpp
@@ -3,13 +3,14 @@
  * See LICENSE in top-level directory.
  */
 
-#include "pallas/pallas_attribute.h"
+#include <cinttypes>
+
 #include "pallas/pallas.h"
+#include "pallas/pallas_attribute.h"
 #include "pallas/pallas_archive.h"
-#include "pallas/pallas_log.h"
 #include "pallas/pallas_read.h"
 
-#include <cinttypes>
+#include "pallas/utils/pallas_log.h"
 
 
 namespace pallas {
