+++ test/sequence_durations.cpp
@@ -4,10 +4,14 @@
  *
  * This is a test for the computation of durations during sequence creation.
  */
+
+#include <iostream>
+
 #include "pallas/pallas.h"
-#include "pallas/pallas_log.h"
 #include "pallas/pallas_record.h"
 #include "pallas/pallas_write.h"
+
+#include "pallas/utils/pallas_log.h"
 
 using namespace pallas;
